requirements.txt
requests
openai
azure-identity